Auth page init
